more refactoring - merge json/jsonx formatters
