+++ audit/event_formatter.go
@@ -1,0 +1,467 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
+package audit
+
+import (
+	"context"
+	"crypto/tls"
+	"errors"
+	"fmt"
+	"time"
+
+	"github.com/jefferai/jsonx"
+
+	"github.com/hashicorp/vault/helper/namespace"
+	"github.com/hashicorp/vault/sdk/logical"
+
+	"github.com/hashicorp/vault/internal/observability/event"
+
+	"github.com/hashicorp/vault/sdk/helper/jsonutil"
+
+	"github.com/hashicorp/eventlogger"
+)
+
+var (
+	_ Formatter        = (*EventFormatter)(nil)
+	_ eventlogger.Node = (*EventFormatter)(nil)
+)
+
+// NewEventFormatter should be used to create an EventFormatter.
+func NewEventFormatter(config FormatterConfig, salter Salter) (*EventFormatter, error) {
+	const op = "audit.NewEventFormatter"
+
+	if salter == nil {
+		return nil, fmt.Errorf("%s: cannot create a new audit formatter with nil salter: %w", op, event.ErrInvalidParameter)
+	}
+
+	return &EventFormatter{
+		salter: salter,
+		config: config,
+	}, nil
+}
+
+// Reopen is a no-op for a formatter node.
+func (_ *EventFormatter) Reopen() error {
+	return nil
+}
+
+// Type describes the type of this node (formatter).
+func (_ *EventFormatter) Type() eventlogger.NodeType {
+	return eventlogger.NodeTypeFormatter
+}
+
+// Process will attempt to parse the incoming event data into a corresponding
+// audit request/response entry which is serialized to JSON and stored within the event.
+func (f *EventFormatter) Process(ctx context.Context, e *eventlogger.Event) (*eventlogger.Event, error) {
+	const op = "audit.(EventFormatter).Process"
+
+	select {
+	case <-ctx.Done():
+		return nil, ctx.Err()
+	default:
+	}
+
+	if e == nil {
+		return nil, fmt.Errorf("%s: event is nil: %w", op, event.ErrInvalidParameter)
+	}
+
+	a, ok := e.Payload.(*auditEvent)
+	if !ok {
+		return nil, fmt.Errorf("%s: cannot parse event payload: %w", op, event.ErrInvalidParameter)
+	}
+
+	var jsonBytes []byte
+
+	switch a.Subtype {
+	case RequestType:
+		entry, err := f.FormatRequest(ctx, a.Data)
+		if err != nil {
+			return nil, fmt.Errorf("%s: unable to parse request from audit event: %w", op, err)
+		}
+
+		jsonBytes, err = jsonutil.EncodeJSON(entry)
+		if err != nil {
+			return nil, fmt.Errorf("%s: unable to format request: %w", op, err)
+		}
+	case ResponseType:
+		entry, err := f.FormatResponse(ctx, a.Data)
+		if err != nil {
+			return nil, fmt.Errorf("%s: unable to parse response from audit event: %w", op, err)
+		}
+
+		jsonBytes, err = jsonutil.EncodeJSON(entry)
+		if err != nil {
+			return nil, fmt.Errorf("%s: unable to format response: %w", op, err)
+		}
+	default:
+		return nil, fmt.Errorf("%s: unknown audit event subtype: %q", op, a.Subtype)
+	}
+
+	// Store the JSON format.
+	e.FormattedAs(JSONFormat.String(), jsonBytes)
+
+	if f.config.RequiredFormat == JSONxFormat {
+
+		xmlBytes, err := jsonx.EncodeJSONBytes(jsonBytes)
+		if err != nil {
+			return nil, fmt.Errorf("%s: unable to encode JSONx using JSON data: %w", op, err)
+		}
+		if xmlBytes == nil {
+			return nil, fmt.Errorf("%s: encoded JSONx was nil: %w", op, err)
+		}
+
+		// Store the JSONx format.
+		e.FormattedAs(JSONxFormat.String(), xmlBytes)
+	}
+
+	return e, nil
+}
+
+// FormatRequest attempts to format the specified logical.LogInput into an RequestEntry.
+func (f *EventFormatter) FormatRequest(ctx context.Context, in *logical.LogInput) (*RequestEntry, error) {
+	switch {
+	case in == nil || in.Request == nil:
+		return nil, errors.New("request to request-audit a nil request")
+	case f.salter == nil:
+		return nil, errors.New("salt func not configured")
+	}
+
+	s, err := f.salter.Salt(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("error fetching salt: %w", err)
+	}
+
+	// Set these to the input values at first
+	auth := in.Auth
+	req := in.Request
+	var connState *tls.ConnectionState
+	if auth == nil {
+		auth = new(logical.Auth)
+	}
+
+	if in.Request.Connection != nil && in.Request.Connection.ConnState != nil {
+		connState = in.Request.Connection.ConnState
+	}
+
+	if !f.config.Raw {
+		auth, err = HashAuth(s, auth, f.config.HMACAccessor)
+		if err != nil {
+			return nil, err
+		}
+
+		req, err = HashRequest(s, req, f.config.HMACAccessor, in.NonHMACReqDataKeys)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	var errString string
+	if in.OuterErr != nil {
+		errString = in.OuterErr.Error()
+	}
+
+	ns, err := namespace.FromContext(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	reqType := in.Type
+	if reqType == "" {
+		reqType = "request"
+	}
+	reqEntry := &RequestEntry{
+		Type:          reqType,
+		Error:         errString,
+		ForwardedFrom: req.ForwardedFrom,
+		Auth: &AuditAuth{
+			ClientToken:               auth.ClientToken,
+			Accessor:                  auth.Accessor,
+			DisplayName:               auth.DisplayName,
+			Policies:                  auth.Policies,
+			TokenPolicies:             auth.TokenPolicies,
+			IdentityPolicies:          auth.IdentityPolicies,
+			ExternalNamespacePolicies: auth.ExternalNamespacePolicies,
+			NoDefaultPolicy:           auth.NoDefaultPolicy,
+			Metadata:                  auth.Metadata,
+			EntityID:                  auth.EntityID,
+			RemainingUses:             req.ClientTokenRemainingUses,
+			TokenType:                 auth.TokenType.String(),
+			TokenTTL:                  int64(auth.TTL.Seconds()),
+		},
+
+		Request: &Request{
+			ID:                    req.ID,
+			ClientID:              req.ClientID,
+			ClientToken:           req.ClientToken,
+			ClientTokenAccessor:   req.ClientTokenAccessor,
+			Operation:             req.Operation,
+			MountPoint:            req.MountPoint,
+			MountType:             req.MountType,
+			MountAccessor:         req.MountAccessor,
+			MountRunningVersion:   req.MountRunningVersion(),
+			MountRunningSha256:    req.MountRunningSha256(),
+			MountIsExternalPlugin: req.MountIsExternalPlugin(),
+			MountClass:            req.MountClass(),
+			Namespace: &Namespace{
+				ID:   ns.ID,
+				Path: ns.Path,
+			},
+			Path:                          req.Path,
+			Data:                          req.Data,
+			PolicyOverride:                req.PolicyOverride,
+			RemoteAddr:                    getRemoteAddr(req),
+			RemotePort:                    getRemotePort(req),
+			ReplicationCluster:            req.ReplicationCluster,
+			Headers:                       req.Headers,
+			ClientCertificateSerialNumber: getClientCertificateSerialNumber(connState),
+		},
+	}
+
+	if !auth.IssueTime.IsZero() {
+		reqEntry.Auth.TokenIssueTime = auth.IssueTime.Format(time.RFC3339)
+	}
+
+	if auth.PolicyResults != nil {
+		reqEntry.Auth.PolicyResults = &PolicyResults{
+			Allowed: auth.PolicyResults.Allowed,
+		}
+
+		for _, p := range auth.PolicyResults.GrantingPolicies {
+			reqEntry.Auth.PolicyResults.GrantingPolicies = append(reqEntry.Auth.PolicyResults.GrantingPolicies, PolicyInfo{
+				Name:          p.Name,
+				NamespaceId:   p.NamespaceId,
+				NamespacePath: p.NamespacePath,
+				Type:          p.Type,
+			})
+		}
+	}
+
+	if req.WrapInfo != nil {
+		reqEntry.Request.WrapTTL = int(req.WrapInfo.TTL / time.Second)
+	}
+
+	if !f.config.OmitTime {
+		reqEntry.Time = time.Now().UTC().Format(time.RFC3339Nano)
+	}
+
+	return reqEntry, nil
+}
+
+// FormatResponse attempts to format the specified logical.LogInput into an ResponseEntry.
+func (f *EventFormatter) FormatResponse(ctx context.Context, in *logical.LogInput) (*ResponseEntry, error) {
+	switch {
+	case in == nil || in.Request == nil:
+		return nil, errors.New("request to response-audit a nil request")
+	case f.salter == nil:
+		return nil, errors.New("salt func not configured")
+	}
+
+	s, err := f.salter.Salt(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("error fetching salt: %w", err)
+	}
+
+	// Set these to the input values at first
+	auth, req, resp := in.Auth, in.Request, in.Response
+	if auth == nil {
+		auth = new(logical.Auth)
+	}
+	if resp == nil {
+		resp = new(logical.Response)
+	}
+	var connState *tls.ConnectionState
+
+	if in.Request.Connection != nil && in.Request.Connection.ConnState != nil {
+		connState = in.Request.Connection.ConnState
+	}
+
+	elideListResponseData := f.config.ElideListResponses && req.Operation == logical.ListOperation
+
+	var respData map[string]interface{}
+	if f.config.Raw {
+		// In the non-raw case, elision of list response data occurs inside HashResponse, to avoid redundant deep
+		// copies and hashing of data only to elide it later. In the raw case, we need to do it here.
+		if elideListResponseData && resp.Data != nil {
+			// Copy the data map before making changes, but we only need to go one level deep in this case
+			respData = make(map[string]interface{}, len(resp.Data))
+			for k, v := range resp.Data {
+				respData[k] = v
+			}
+
+			doElideListResponseData(respData)
+		} else {
+			respData = resp.Data
+		}
+	} else {
+		auth, err = HashAuth(s, auth, f.config.HMACAccessor)
+		if err != nil {
+			return nil, err
+		}
+
+		req, err = HashRequest(s, req, f.config.HMACAccessor, in.NonHMACReqDataKeys)
+		if err != nil {
+			return nil, err
+		}
+
+		resp, err = HashResponse(s, resp, f.config.HMACAccessor, in.NonHMACRespDataKeys, elideListResponseData)
+		if err != nil {
+			return nil, err
+		}
+
+		respData = resp.Data
+	}
+
+	var errString string
+	if in.OuterErr != nil {
+		errString = in.OuterErr.Error()
+	}
+
+	ns, err := namespace.FromContext(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	var respAuth *AuditAuth
+	if resp.Auth != nil {
+		respAuth = &AuditAuth{
+			ClientToken:               resp.Auth.ClientToken,
+			Accessor:                  resp.Auth.Accessor,
+			DisplayName:               resp.Auth.DisplayName,
+			Policies:                  resp.Auth.Policies,
+			TokenPolicies:             resp.Auth.TokenPolicies,
+			IdentityPolicies:          resp.Auth.IdentityPolicies,
+			ExternalNamespacePolicies: resp.Auth.ExternalNamespacePolicies,
+			NoDefaultPolicy:           resp.Auth.NoDefaultPolicy,
+			Metadata:                  resp.Auth.Metadata,
+			NumUses:                   resp.Auth.NumUses,
+			EntityID:                  resp.Auth.EntityID,
+			TokenType:                 resp.Auth.TokenType.String(),
+			TokenTTL:                  int64(resp.Auth.TTL.Seconds()),
+		}
+		if !resp.Auth.IssueTime.IsZero() {
+			respAuth.TokenIssueTime = resp.Auth.IssueTime.Format(time.RFC3339)
+		}
+	}
+
+	var respSecret *Secret
+	if resp.Secret != nil {
+		respSecret = &Secret{
+			LeaseID: resp.Secret.LeaseID,
+		}
+	}
+
+	var respWrapInfo *ResponseWrapInfo
+	if resp.WrapInfo != nil {
+		token := resp.WrapInfo.Token
+		if jwtToken := parseVaultTokenFromJWT(token); jwtToken != nil {
+			token = *jwtToken
+		}
+		respWrapInfo = &ResponseWrapInfo{
+			TTL:             int(resp.WrapInfo.TTL / time.Second),
+			Token:           token,
+			Accessor:        resp.WrapInfo.Accessor,
+			CreationTime:    resp.WrapInfo.CreationTime.UTC().Format(time.RFC3339Nano),
+			CreationPath:    resp.WrapInfo.CreationPath,
+			WrappedAccessor: resp.WrapInfo.WrappedAccessor,
+		}
+	}
+
+	respType := in.Type
+	if respType == "" {
+		respType = "response"
+	}
+	respEntry := &ResponseEntry{
+		Type:      respType,
+		Error:     errString,
+		Forwarded: req.ForwardedFrom != "",
+		Auth: &AuditAuth{
+			ClientToken:               auth.ClientToken,
+			Accessor:                  auth.Accessor,
+			DisplayName:               auth.DisplayName,
+			Policies:                  auth.Policies,
+			TokenPolicies:             auth.TokenPolicies,
+			IdentityPolicies:          auth.IdentityPolicies,
+			ExternalNamespacePolicies: auth.ExternalNamespacePolicies,
+			NoDefaultPolicy:           auth.NoDefaultPolicy,
+			Metadata:                  auth.Metadata,
+			RemainingUses:             req.ClientTokenRemainingUses,
+			EntityID:                  auth.EntityID,
+			EntityCreated:             auth.EntityCreated,
+			TokenType:                 auth.TokenType.String(),
+			TokenTTL:                  int64(auth.TTL.Seconds()),
+		},
+
+		Request: &Request{
+			ID:                    req.ID,
+			ClientToken:           req.ClientToken,
+			ClientTokenAccessor:   req.ClientTokenAccessor,
+			ClientID:              req.ClientID,
+			Operation:             req.Operation,
+			MountPoint:            req.MountPoint,
+			MountType:             req.MountType,
+			MountAccessor:         req.MountAccessor,
+			MountRunningVersion:   req.MountRunningVersion(),
+			MountRunningSha256:    req.MountRunningSha256(),
+			MountIsExternalPlugin: req.MountIsExternalPlugin(),
+			MountClass:            req.MountClass(),
+			Namespace: &Namespace{
+				ID:   ns.ID,
+				Path: ns.Path,
+			},
+			Path:                          req.Path,
+			Data:                          req.Data,
+			PolicyOverride:                req.PolicyOverride,
+			RemoteAddr:                    getRemoteAddr(req),
+			RemotePort:                    getRemotePort(req),
+			ClientCertificateSerialNumber: getClientCertificateSerialNumber(connState),
+			ReplicationCluster:            req.ReplicationCluster,
+			Headers:                       req.Headers,
+		},
+
+		Response: &Response{
+			MountPoint:            req.MountPoint,
+			MountType:             req.MountType,
+			MountAccessor:         req.MountAccessor,
+			MountRunningVersion:   req.MountRunningVersion(),
+			MountRunningSha256:    req.MountRunningSha256(),
+			MountIsExternalPlugin: req.MountIsExternalPlugin(),
+			MountClass:            req.MountClass(),
+			Auth:                  respAuth,
+			Secret:                respSecret,
+			Data:                  respData,
+			Warnings:              resp.Warnings,
+			Redirect:              resp.Redirect,
+			WrapInfo:              respWrapInfo,
+			Headers:               resp.Headers,
+		},
+	}
+
+	if auth.PolicyResults != nil {
+		respEntry.Auth.PolicyResults = &PolicyResults{
+			Allowed: auth.PolicyResults.Allowed,
+		}
+
+		for _, p := range auth.PolicyResults.GrantingPolicies {
+			respEntry.Auth.PolicyResults.GrantingPolicies = append(respEntry.Auth.PolicyResults.GrantingPolicies, PolicyInfo{
+				Name:          p.Name,
+				NamespaceId:   p.NamespaceId,
+				NamespacePath: p.NamespacePath,
+				Type:          p.Type,
+			})
+		}
+	}
+
+	if !auth.IssueTime.IsZero() {
+		respEntry.Auth.TokenIssueTime = auth.IssueTime.Format(time.RFC3339)
+	}
+	if req.WrapInfo != nil {
+		respEntry.Request.WrapTTL = int(req.WrapInfo.TTL / time.Second)
+	}
+
+	if !f.config.OmitTime {
+		respEntry.Time = time.Now().UTC().Format(time.RFC3339Nano)
+	}
+
+	return respEntry, nil
+}
