+++ builtin/audit/file/backend.go
@@ -125,7 +125,7 @@ func Factory(ctx context.Context, conf *audit.BackendConfig) (audit.Backend, err
 	b.salt.Store((*salt.Salt)(nil))
 
 	// Configure the formatter for either case.
-	f, err := audit.NewAuditFormatter(b)
+	f, err := audit.NewEventFormatter(b)
 	if err != nil {
 		return nil, fmt.Errorf("error creating formatter: %w", err)
 	}
@@ -137,7 +137,7 @@ func Factory(ctx context.Context, conf *audit.BackendConfig) (audit.Backend, err
 		w = &audit.JSONxWriter{Prefix: conf.Config["prefix"]}
 	}
 
-	fw, err := audit.NewAuditFormatterWriter(f, w)
+	fw, err := audit.NewEventFormatterWriter(f, w)
 	if err != nil {
 		return nil, fmt.Errorf("error creating formatter writer: %w", err)
 	}
@@ -166,7 +166,7 @@ func Factory(ctx context.Context, conf *audit.BackendConfig) (audit.Backend, err
 type Backend struct {
 	path string
 
-	formatter    *audit.AuditFormatterWriter
+	formatter    *audit.EventFormatterWriter
 	formatConfig audit.FormatterConfig
 
 	fileLock sync.RWMutex
